Create a new absorber description interface fields::absorber::Absorber
This interface is to be used for getting absorber size and string properties.
Move the previously created fields/absorber/NumCells.hpp into this file.
This does not provide the unified computational interface, to be done in following PRs.
Fixes #3204.
