+++ include/picongpu/plugins/adios/WriteMeta.hpp
@@ -24,7 +24,7 @@
 #include "picongpu/plugins/common/stringHelpers.hpp"
 #include <pmacc/Environment.hpp>
 
-#include "picongpu/fields/absorber/ExponentialDamping.hpp"
+#include "picongpu/fields/absorber/Absorber.hpp"
 #include "picongpu/fields/currentInterpolation/CurrentInterpolation.hpp"
 
 #include "picongpu/traits/SIBaseUnits.hpp"
@@ -189,7 +189,7 @@ namespace writeMeta
              *    3D: z-lower, z-upper, y-lower, y-upper, x-lower, x-upper
              *    2D: y-lower, y-upper, x-lower, x-upper
              */
-            GetStringProperties<fields::absorber::ExponentialDamping> fieldBoundaryProp;
+            GetStringProperties<fields::absorber::Absorber> fieldBoundaryProp;
             std::list<std::string> listFieldBoundary;
             std::list<std::string> listFieldBoundaryParam;
             for( uint32_t i = NumberOfExchanges<simDim>::value - 1; i > 0; --i )
